Corrections related to the ToolBar class
- removed unnecessary dependencies: ImageDisplayer and ResliceDisplayer
- created an event group to replace usage of 'ugly' dependencies

+++ src/main/java/org/example/mydicomviewer/views/WindowingPopup.java
@@ -20,11 +20,11 @@ public class WindowingPopup extends JPanel {
         add(width);
     }
 
-    public double getWindowCenter() {
-        return Double.parseDouble(center.getText());
+    public int getWindowCenter() {
+        return Integer.parseInt(center.getText());
     }
 
-    public double getWindowWidth() {
-        return Double.parseDouble(width.getText());
+    public int getWindowWidth() {
+        return Integer.parseInt(width.getText());
     }
 }
